Fix streets gatekeeper - issue 26
The gatekeeper's target coordinates were 70,256 but it couldn't get further
south than 70,250, so it would never reach its target. This changes
its target to 70,250. This also includes a hack to correct the target
Y coordinate (the goToY variable) in case an older saved game was loaded,
deserializing an old value for goToY.